refactor(test): address PR #1385 KP13 review (Boy-Scout + 3 docstring fixes)
Single bundled change addressing KP13's CHANGES_REQUESTED review on
PR #1385. Issue #1357's mission is "branch-coverage parity across all
_classify_by_message branches" — leaving the schema/auth parametrize
tuples in test_tool_error_signaling.py was unfinished mission, not
out-of-scope follow-up.

**Boy Scout — schema/auth migration**

Moved from test_tool_error_signaling.py::TestSchemaAndAuthClassification
into test_helper_classifier.py::TestClassifyByMessageBranches:

- SCHEMA_MARKER_MESSAGES (13 tuples) + test_schema_marker_classified_as_validation_failed
- AUTH_PHRASE_MESSAGES (4 tuples) + test_auth_phrase_classified
- test_401_status_still_classified_as_auth

TestClassifyByMessageBranches now covers all 9 elif branches in
_classify_by_message — full per-file mutation-grade pin.

Kept at TestSchemaAndAuthClassification (orthogonal concerns, not
classifier-branch parametrize):

- test_schema_phrase_without_command_prefix_not_validation
  (Command failed: outer gate)
- test_authorized_keys_substring_not_auth_error (#993 root-cause guard)
- test_command_error_unknown_message_is_service_call_failed
  (HomeAssistantCommandError typed-dispatch fallback)
- test_auth_required_handshake_is_connection_error /
  test_typed_auth_error_classified_as_auth /
  test_typed_connection_error_classified_as_connection
  (typed-exception dispatch via _classify_exception)

Class docstrings updated on both sides to reflect the new split:
TestClassifyByMessageBranches lists all 9 branches now covered;
TestSchemaAndAuthClassification narrowed to precedence/gate +
false-positive guard + typed-dispatch carve-out.

**While-you're-in (3 docstring fixes)**

- test_plain_not_found_with_entity_id_context_promotes_to_entity_not_found
  docstring: the "drop the not-found clause" failure mode was wrong.
  With the clause removed the input matches no other arm and drops to
  INTERNAL_ERROR via the final else, not RESOURCE_NOT_FOUND.
- Line-anchor citations switched to symbol-level refs:
  - "helpers.py:258-327" → "_classify_by_message" (function name is
    the durable handle)
  - "test_tool_error_signaling.py:229" →
    "::TestErrorCodeMapping::test_connection_error_in_message_maps_correctly"
- "pre-#1345 substring" reworded to "the long-standing 'not found'
  substring (predates #1345's 'unknown config specified' addition)".

2869 unit tests pass (1 numpy-tripwire skip on Alpine, expected).
